Cross-platform support, help option, misc changes to runner Improve cross-platform support for runner.py
This should now function on Windows, Linux and Mac, looking in the
default installation location for the platform.

Remove runner/config.py

Add useful help support to the IDF5.run function

This uses a decorator to pass through the help from the wrapped
function. It serves as a replacement for the EnergyPlus CLI `help`
flag since we want to avoid stepping on the python built-in `help`.

PEP8 spacing

Change runlist to joblist and improve num CPU settings