Rewrite the the "no warning" test to avoid the new pytest behaviour
This error was caused by pytest raising an error when seeing that no
warning was raised during a test using `pytest.warn`. For some reason
this wasn't being caught by `pytest.raises`. In any case, the test
wasn't very readable so we have rewritten it to make it clearer how it
work.

Fixes #163